move duplicated selectors to slice
