+++ apps/fishing-map/features/analysis/Analysis.tsx
@@ -10,7 +10,7 @@ import { Dataset, DatasetTypes } from '@globalfishingwatch/api-types'
 import { useFeatureState, useLoginRedirect } from '@globalfishingwatch/react-hooks'
 import { useLocationConnect } from 'routes/routes.hook'
 import sectionStyles from 'features/workspace/shared/Sections.module.css'
-import { selectUserData, isGFWUser } from 'features/user/user.slice'
+import { selectUserData, isGFWUser, isGuestUser } from 'features/user/user.slice'
 import { AsyncReducerStatus } from 'utils/async-slice'
 import useMapInstance from 'features/map/map-context.hooks'
 import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
@@ -20,7 +20,6 @@ import {
 } from 'features/dataviews/dataviews.selectors'
 import { getRelatedDatasetByType } from 'features/datasets/datasets.selectors'
 import { getActivityFilters, getEventLabel } from 'utils/analytics'
-import { isGuestUser } from 'features/selectors/user.selectors'
 import { selectAnalysisQuery, selectAnalysisTypeQuery } from 'features/app/app.selectors'
 import { WorkspaceAnalysisType } from 'types'
 import { useMapFitBounds } from 'features/map/map-viewport.hooks'
