+++ apps/fishing-map/features/map/Map.tsx
@@ -35,7 +35,7 @@ import { getEventLabel } from 'utils/analytics'
 import { selectIsAnalyzing, selectShowTimeComparison } from 'features/analysis/analysis.selectors'
 import Hint from 'features/help/hints/Hint'
 import { isWorkspaceLocation } from 'routes/routes.selectors'
-import { selectDataviewInstancesResolved } from 'features/selectors/dataviews.selectors'
+import { selectDataviewInstancesResolved } from 'features/dataviews/dataviews.slice'
 import { useMapLoaded } from 'features/map/map-style.hooks'
 import PopupWrapper from './popups/PopupWrapper'
 import useViewport, { useMapBounds } from './map-viewport.hooks'
