+++ apps/fishing-map/features/sidebar/CategoryTabs.tsx
@@ -8,8 +8,7 @@ import { useFeatureState } from '@globalfishingwatch/react-hooks'
 import { WorkspaceCategories } from 'data/workspaces'
 import { HOME, USER, WORKSPACES_LIST } from 'routes/routes'
 import { selectLocationCategory, selectLocationType } from 'routes/routes.selectors'
-import { selectUserData } from 'features/user/user.slice'
-import { isGuestUser } from 'features/selectors/user.selectors'
+import { selectUserData, isGuestUser } from 'features/user/user.slice'
 import { useClickedEventConnect } from 'features/map/map.hooks'
 import useMapInstance from 'features/map/map-context.hooks'
 import { selectAvailableWorkspacesCategories } from 'features/workspaces-list/workspaces-list.selectors'
