+++ apps/fishing-map/features/user/user.slice.ts
@@ -1,4 +1,4 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit'
+import { createSlice, createAsyncThunk, createSelector } from '@reduxjs/toolkit'
 import {
   GFWAPI,
   getAccessTokenFromUrl,
@@ -92,5 +92,9 @@ export const selectUserData = (state: RootState) => state.user.data
 export const selectUserStatus = (state: RootState) => state.user.status
 export const selectUserLogged = (state: RootState) => state.user.logged
 export const isGFWUser = (state: RootState) => state.user.data?.groups.includes(GFW_GROUP_ID)
+
+export const isGuestUser = createSelector([selectUserData], (userData) => {
+  return userData?.type === GUEST_USER_TYPE
+})
 
 export default userSlice.reducer
