+++ apps/fishing-map/features/workspace/environmental/EnvironmentalLayerPanel.tsx
@@ -9,7 +9,7 @@ import styles from 'features/workspace/shared/LayerPanel.module.css'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
 import { selectUserId } from 'features/user/user.selectors'
 import { useAutoRefreshImportingDataset } from 'features/datasets/datasets.hook'
-import { isGuestUser } from 'features/selectors/user.selectors'
+import { isGuestUser } from 'features/user/user.slice'
 import DatasetNotFound from '../shared/DatasetNotFound'
 import Color from '../common/Color'
 import LayerSwitch from '../common/LayerSwitch'
