+++ apps/fishing-map/features/workspace/shared/NewWorkspaceModal.tsx
@@ -17,7 +17,7 @@ import { selectViewport } from 'features/app/app.selectors'
 import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
 import { getDatasetsInDataviews } from 'features/datasets/datasets.utils'
 import { PRIVATE_SUFIX, PUBLIC_SUFIX } from 'data/config'
-import { selectDataviewInstancesMerged } from 'features/selectors/dataviews.selectors'
+import { selectDataviewInstancesMerged } from 'features/dataviews/dataviews.slice'
 import { selectUserData } from 'features/user/user.slice'
 import { selectUserWorkspaceEditPermissions } from 'features/user/user.selectors'
 import { selectWorkspaceId } from 'routes/routes.selectors'
