+++ apps/fishing-map/features/workspace/vessels/VesselLayerPanel.tsx
@@ -21,12 +21,11 @@ import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
 import { selectResourceByUrl } from 'features/resources/resources.slice'
 import { VESSEL_DATAVIEW_INSTANCE_PREFIX } from 'features/dataviews/dataviews.utils'
 import ExpandedContainer from 'features/workspace/shared/ExpandedContainer'
-import { isGuestUser } from 'features/selectors/user.selectors'
+import { isGuestUser, isGFWUser } from 'features/user/user.slice'
 import I18nDate from 'features/i18n/i18nDate'
 import I18nFlag from 'features/i18n/i18nFlag'
 import { getDatasetLabel } from 'features/datasets/datasets.utils'
 import { setDownloadTrackVessel } from 'features/download/downloadTrack.slice'
-import { isGFWUser } from 'features/user/user.slice'
 import LocalStorageLoginLink from 'routes/LoginLink'
 import Color from '../common/Color'
 import LayerSwitch from '../common/LayerSwitch'
