+++ apps/fishing-map/features/workspace/vessels/VesselsSection.tsx
@@ -5,7 +5,7 @@ import { event as uaEvent } from 'react-ga'
 import { useTranslation } from 'react-i18next'
 import { IconButton } from '@globalfishingwatch/ui-components'
 import { useLocationConnect } from 'routes/routes.hook'
-import { selectVesselsDataviews } from 'features/selectors/dataviews.selectors'
+import { selectVesselsDataviews } from 'features/dataviews/dataviews.slice'
 import styles from 'features/workspace/shared/Sections.module.css'
 import { isBasicSearchAllowed } from 'features/search/search.selectors'
 import VesselEventsLegend from './VesselEventsLegend'
