+++ apps/fishing-map/next.config.js
@@ -37,9 +37,10 @@ const nextConfig = {
     // config.optimization.minimize = false
     // config.plugins.push(
     //   new CircularDependencyPlugin({
+    //     // exclude detection of files based on a RegExp
     //     exclude: /node_modules/,
     //     // add errors to webpack instead of warnings
-    //     failOnError: false,
+    //     failOnError: true,
     //     // allow import cycles that include an asyncronous import,
     //     // e.g. via import(/* webpackMode: "weak" */ './file.js')
     //     allowAsyncCycles: true,
